refactor: styles refactor
