+++ src/components/Calculator/config/pages.js
