+++ src/constants/index.js
@@ -1,6 +1,6 @@
+export * from '../components/Navigation/config/links';
 export * from './buttons';
 export * from './componentsConstants';
 export * from './fakeLanguagesData';
 export * from './fakeThemeData';
-export * from './links';
 export * from './sizes';
