+++ src/styles/applicationThemes.js
@@ -27,10 +27,3 @@ export const themeOptions = {
         ...sharedOptions,
     },
 };
-
-export const activeStyle = {
-    borderBottom: `2px solid #fefefe`,
-};
-export const nonActive = {
-    opacity: '0.7',
-};
