+++ src/styles/sharedOptions.js
@@ -1,7 +1,8 @@
 const sharedOptions = {
     top: {
-        30: 30,
         10: 10,
+        15: 15,
+        30: 30,
         50: 50,
     },
     right: {
@@ -17,70 +18,70 @@ const sharedOptions = {
         4: 4,
     },
     gap: {
-        32: 32,
-        15: 15,
         0: 0,
+        15: 15,
+        32: 32,
     },
     widthOptions: {
-        100: 100,
-        1920: 1920,
-        400: 400,
-        250: 250,
-        500: 500,
-        355: 355,
-        90: 90,
-        50: 50,
-        40: 40,
-        18: 18,
-        17: 17,
         16: 16,
+        17: 17,
+        18: 18,
         20: 20,
+        40: 40,
+        50: 50,
+        90: 90,
+        100: 100,
+        250: 250,
+        355: 355,
+        400: 400,
+        500: 500,
+        1920: 1920,
     },
     heightOptions: {
-        100: 100,
-        1080: 1080,
-        250: 250,
-        960: 960,
         0: 0,
-        840: 840,
-        670: 670,
-        55: 55,
-        90: 90,
-        50: 50,
-        49: 49,
-        158: 158,
-        80: 80,
-        40: 40,
         2: 2,
-        45: 45,
-        60: 60,
-        900: 900,
-        730: 730,
         30: 30,
-        800: 800,
-        125: 125,
+        40: 40,
+        45: 45,
+        49: 49,
+        50: 50,
+        55: 55,
+        60: 60,
+        80: 80,
+        90: 90,
+        100: 100,
         110: 110,
+        125: 125,
+        158: 158,
+        250: 250,
+        670: 670,
+        730: 730,
+        800: 800,
+        840: 840,
+        900: 900,
+        960: 960,
+        1080: 1080,
     },
     margin: {
         0: 0,
-        475: 475,
-        390: 390,
+        10: 10,
+        20: 20,
+        32: 32,
         90: 90,
+        390: 390,
+        475: 475,
         center: '0 auto',
         header: '20px 0px 0px 30px',
         navigation: '20px 30px 0px 0px',
         burger: '30px 30px 0px 0px',
-        32: 32,
-        20: 20,
-        10: 10,
     },
     padding: {
         0: 0,
-        30: 30,
-        20: 20,
-        25: 25,
         5: 5,
         10: 10,
+        20: 20,
+        25: 25,
+        30: 30,
         displayPadding: '0px 110px 0px 60px',
         displayLaptopPadding: '0px 50px',
         displayTabletPadding: '0px 30px',
@@ -91,11 +92,11 @@ const sharedOptions = {
         autoOverflow: 'auto',
     },
     fontSize: {
-        xxl: 64,
-        xl: 46,
-        l: 32,
-        m: 28,
         s: 24,
+        m: 28,
+        l: 32,
+        xl: 46,
+        xxl: 64,
     },
     fontWeight: {
         700: 700,
@@ -107,14 +108,6 @@ const sharedOptions = {
     borderRadius: {
         default: '10px',
     },
-    cursor: {
-        pointer: 'pointer',
-        noToday: 'not-allowed',
-    },
-    textAlign: {
-        left: 'left',
-        center: 'center',
-    },
     background: {
         white: '#fff',
         darkGrey: '#707070',
@@ -122,6 +115,7 @@ const sharedOptions = {
     color: {
         white: '#fff',
         black: '#000',
+        beige: '#fefefe',
     },
     transition: {
         default: 'all .3s ease-out',
@@ -130,7 +124,9 @@ const sharedOptions = {
         40: 40,
     },
     opacity: {
+        1: 1,
         default: '0.3',
+        navigation: '0.7',
     },
 };
